Moving awkward._connect.pyarrow into a package
