+++ src/awkward/operations/ak_to_parquet.py
@@ -8,7 +8,7 @@ from os import fsdecode
 import fsspec
 
 import awkward as ak
-from awkward._connect.pyarrow_table_conv import convert_awkward_arrow_table_to_native
+from awkward._connect.pyarrow.table_conv import convert_awkward_arrow_table_to_native
 from awkward._dispatch import high_level_function
 from awkward._nplikes.numpy_like import NumpyMetadata
